Better test script

+++ website/play/test.sh
@@ -2,7 +2,12 @@
 
 set -euo pipefail
 
-curl http://localhost:3000/build --json @test.json --output out.gba.gz
+if [ "$#" != 1 ]; then
+    echo "Usage: $0 <rust file to compile>"
+    exit 1
+fi
+
+curl -v http://localhost:3000/build --json "$(jq -Rs '{code: .}' < "$1")" --output out.gba.gz
 if gunzip out.gba.gz; then
     mgba-qt out.gba
     rm -f out.gba out.gba.sav
